+++ apps/testapp-embedded/src/app.d.ts
@@ -14,6 +14,7 @@ declare module 'virtual:sdocs' {
 	import type { DocEntry } from 'sdocs';
 	export const docs: DocEntry[];
 	export const cssNames: string[];
+	export const pageModules: Record<string, () => Promise<{ default: unknown }>>;
 	export default docs;
 }